finished and displayed BobsEpisodesList
